elegacy-backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-pymongo
flask-bcrypt
cryptography
python-dotenv
werkzeug
